Updated setup script

+++ setup.sh
@@ -14,48 +14,55 @@ function link
   target=$PWD/$file
   link=~/.$file
 
-  #
-  # Check if argument file does not exist
-  #
-  if [ ! -e $file ]; then
-    echo "File does not exist: $file"
-    exit 1
+  if [ "$safemode" -eq 1 ]; then
+    echo " $link"
+  else
+    if [ ! -e $file ]; then
+      echo "Error: File does not exist: $file"
+      exit 1
+    fi
+    echo "  ~/.$file"
+    remove_link $link
+    safe_remove_file $link
+    create_directory $dir
+    ln -s $target $link
   fi
-
-  #
-  # Initial output
-  #
-  echo "Linking: ~/.$file"
-
-  #
-  # Create directory if needed
-  #
-  [ ! "$dir" = "." ] && [ ! -d "$dir" ] && mkdir -p $dir
-
-  #
-  # Check if file is already present
-  #
-  [ -L $link ] && remove_link $link
-  [ -e $link ] && safe_remove_file $link
-
-  #
-  # Create link
-  #
-  ln -s $target $link
 }
 function remove_link
 {
-  echo "  -- Removed old link"
-  rm $1
+  [ -L $1 ] && rm $1
 }
 function safe_remove_file
 {
   bfile=/tmp/$(basename $1)_$RANDOM
-  echo " -- Backup of existing file: $bfile"
-  mv $1 $bfile
+  if [ -e $1 ]; then
+    mv $1 $bfile
+    file_removed=1
+  fi
+}
+function create_directory
+{
+  if [ ! "$1" = "." ]; then
+    checkdir=$HOME/.$1
+    [ -d "$checkdir" ] && return
+    mkdir -p "$checkdir"
+    dir_created=1
+  fi
 }
 
-files=$(find -type f ! -path "./.*" ! -name "setup.sh")
-for file in $files; do
-  link ${file:2}
+file_removed=0
+dir_created=0
+safemode=0
+if [ "$#" -eq 0 -o "$1" != "-r" ]; then
+  safemode=1
+  echo "Please run the script with option -r to apply changes."
+  echo "This run will only show which files will be linked."
+fi
+
+echo "Linking files"
+for file in $(cat list-of-files.txt); do
+  link ${file}
 done
+
+[ "$file_removed" -eq "1" ] && echo "Note: Some files were moved to /tmp"
+[ "$dir_created"  -eq "1" ] && echo "Note: Some dirs were created"
